Factor PayeeTable out to component file
